backend/requirements.txt
# BorgDash Backend Requirements
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
toml==0.10.2
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
PyJWT==2.10.1
